Added mocking forever support
In the original code, the requests list/queue is processed when new requests are executed. And earlier mocks are removed on-the-fly when hiroaki responds. It is actually not possible to compare and remove previous matchers immediately when they're added since they don't implement `equals` functionality.

With this approach here is some principles I put into place:

- Regular `respond` always takes priority regardless of being added before or after.
- The most recently added `respondForever` takes precedence. As mentioned above, I cannot clear previous matchers immediately when a new one is added. I could potentially clear previous one when a `respondForever` is matched successfully. I figured that this is not a good idea since 1. it is not trivial, 2. there might be cases where previously added matcher can be useful for other requests even though it is matched for one request.